:sparkles: first version of the server
An event bus is used to uncouple the websocket server
implementation from its user interface.
We will begin with an http server with a POST route to push a message on
the websocket clisnts.
Later, we could use a message broker. Then,
the only thing to do will be to add a message broker client that will
publish the message into the event bus.